Run-time errors and issues created have been mostly solved
