+++ Project II/Project II/x64/Debug/Project II.tlog/Cl.items.tlog
@@ -4,4 +4,5 @@ C:\Users\DELL\OneDrive - University of Guelph\Desktop\Conestoga\Winter 25\05 - P
 C:\Users\DELL\OneDrive - University of Guelph\Desktop\Conestoga\Winter 25\05 - Project II\Integration\Project II\menu_customer.c;C:\Users\DELL\OneDrive - University of Guelph\Desktop\Conestoga\Winter 25\05 - Project II\Integration\Project II\Project II\x64\Debug\menu_customer.obj
 C:\Users\DELL\OneDrive - University of Guelph\Desktop\Conestoga\Winter 25\05 - Project II\Integration\Project II\menu_staff.c;C:\Users\DELL\OneDrive - University of Guelph\Desktop\Conestoga\Winter 25\05 - Project II\Integration\Project II\Project II\x64\Debug\menu_staff.obj
 C:\Users\DELL\OneDrive - University of Guelph\Desktop\Conestoga\Winter 25\05 - Project II\Integration\Project II\order.c;C:\Users\DELL\OneDrive - University of Guelph\Desktop\Conestoga\Winter 25\05 - Project II\Integration\Project II\Project II\x64\Debug\order.obj
+C:\Users\DELL\OneDrive - University of Guelph\Desktop\Conestoga\Winter 25\05 - Project II\Integration\Project II\order_flow.c;C:\Users\DELL\OneDrive - University of Guelph\Desktop\Conestoga\Winter 25\05 - Project II\Integration\Project II\Project II\x64\Debug\order_flow.obj
 C:\Users\DELL\OneDrive - University of Guelph\Desktop\Conestoga\Winter 25\05 - Project II\Integration\Project II\welcome.c;C:\Users\DELL\OneDrive - University of Guelph\Desktop\Conestoga\Winter 25\05 - Project II\Integration\Project II\Project II\x64\Debug\welcome.obj
