+++ Project II/main.c
@@ -3,24 +3,31 @@
 #include <string.h>
 #include <time.h>
 
+// Core modules
 #include "menu.h"
 #include "welcome.h"
-#include "billing.h"
 #include "order.h"
+#include "billing.h"
 
 int main(int argc, char* argv[]) {
-    srand((unsigned int)time(NULL)); // Seed random
+    // Seed the random number generator (for discounts, random table assignment, etc.)
+    srand((unsigned int)time(NULL));
 
+    // Check arguments to decide which mode to run
     if (argc == 1) {
-        // No arguments: Customer Version
+        // No arguments: Run as customer
         runCustomerVersion();
     }
     else if (argc == 2 && strcmp(argv[1], "1") == 0) {
-        // Argument "1": Staff Version
+        // Argument "1": Run as staff
         runStaffVersion();
     }
     else {
-        printf("Invalid argument!\nUsage:\n  No argument -> Customer version\n  1 -> Staff version\n");
+        // Invalid usage
+        printf("Invalid argument!\n");
+        printf("Usage:\n");
+        printf("  ./restaurant         -> Run as Customer\n");
+        printf("  ./restaurant 1       -> Run as Staff\n");
         return 1;
     }
