+++ Project II/menu_customer.c
@@ -1,9 +1,6 @@
 #include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
 #include "menu.h"
 
-// Displays the 3-option customer menu
 void DisplayCustomerMenu() {
     printf("\n==== Customer Menu ====\n");
     printf("1. View Menu (displays all menus in order)\n");
