+++ Project II/order.c
@@ -2,7 +2,7 @@
 #include <string.h>
 #include <stdio.h>
 
-// Function to add an item to an order list
+// Add an item to the order
 int addToOrder(Order orders[], int* orderCount, const char itemID[], int quantity, const MenuItem menu[], int menuSize) {
     if (*orderCount >= MAX_ORDERS) {
         printf("Order limit reached. Cannot add more items.\n");
@@ -10,7 +10,6 @@ int addToOrder(Order orders[], int* orderCount, const char itemID[], int quantit
     }
 
     for (int i = 0; i < menuSize; i++) {
-        // Compare integer itemID with formatted string
         char idBuffer[10];
         sprintf_s(idBuffer, sizeof(idBuffer), "%d", menu[i].itemID);
 
@@ -20,7 +19,6 @@ int addToOrder(Order orders[], int* orderCount, const char itemID[], int quantit
             strcpy_s(orders[*orderCount].itemCategory, sizeof(orders[*orderCount].itemCategory), menu[i].itemCategory);
             orders[*orderCount].price = menu[i].itemPrice;
             orders[*orderCount].quantity = quantity;
-
             (*orderCount)++;
             printf("Item '%s' added successfully to order.\n", menu[i].itemName);
             return 1;
@@ -31,21 +29,7 @@ int addToOrder(Order orders[], int* orderCount, const char itemID[], int quantit
     return 0;
 }
 
-// Function to retrieve menu item details based on item ID
-int getMenuItemDetails(const MenuItem menu[], int menuSize, const char itemID[], char* itemName, float* price) {
-    for (int i = 0; i < menuSize; i++) {
-        char idBuffer[10];
-        sprintf_s(idBuffer, sizeof(idBuffer), "%d", menu[i].itemID);
-        if (strcmp(idBuffer, itemID) == 0) {
-            strcpy_s(itemName, 50, menu[i].itemName);
-            *price = menu[i].itemPrice;
-            return 1;
-        }
-    }
-    return 0;
-}
-
-// Function to apply a discount before sending to Billing Module
+// Apply discount (flat 10% off)
 void applyDiscount(Order orders[], int orderCount, char discountChoice) {
     if (discountChoice == 'Y' || discountChoice == 'y') {
         printf("Applying discount to the order...\n");
@@ -61,24 +45,24 @@ void applyDiscount(Order orders[], int orderCount, char discountChoice) {
 
 // Calculate delivery time
 int calculateDeliveryTime(Order orders[], int orderCount) {
-    int time = 10; // Base time
+    int time = 15; // Base delivery time
     for (int i = 0; i < orderCount; i++) {
-        if (strcmp(orders[i].itemCategory, "Drinks") == 0) time += 1;
-        else if (strcmp(orders[i].itemCategory, "Appetizers") == 0) time += 3;
-        else if (strcmp(orders[i].itemCategory, "Main Course") == 0) time += 5;
-        else if (strcmp(orders[i].itemCategory, "Desserts") == 0) time += 2;
+        if (strcmp(orders[i].itemCategory, "Appetizers") == 0) time += orders[i].quantity * 3;
+        else if (strcmp(orders[i].itemCategory, "Main Course") == 0) time += orders[i].quantity * 7;
+        else if (strcmp(orders[i].itemCategory, "Desserts") == 0) time += orders[i].quantity * 2;
+        else if (strcmp(orders[i].itemCategory, "Drinks") == 0) time += orders[i].quantity * 1;
     }
     return time;
 }
 
-// Calculate pickup time
+// Calculate pickup time (excluding delivery buffer)
 int calculatePickupTime(Order orders[], int orderCount) {
     int time = 0;
     for (int i = 0; i < orderCount; i++) {
-        if (strcmp(orders[i].itemCategory, "Drinks") == 0) time += 1;
-        else if (strcmp(orders[i].itemCategory, "Appetizers") == 0) time += 3;
-        else if (strcmp(orders[i].itemCategory, "Main Course") == 0) time += 5;
-        else if (strcmp(orders[i].itemCategory, "Desserts") == 0) time += 2;
+        if (strcmp(orders[i].itemCategory, "Appetizers") == 0) time += orders[i].quantity * 3;
+        else if (strcmp(orders[i].itemCategory, "Main Course") == 0) time += orders[i].quantity * 7;
+        else if (strcmp(orders[i].itemCategory, "Desserts") == 0) time += orders[i].quantity * 2;
+        else if (strcmp(orders[i].itemCategory, "Drinks") == 0) time += orders[i].quantity * 1;
     }
     return time;
 }
