+++ Project II/order.h
@@ -4,19 +4,20 @@
 
 #define MAX_ORDERS 100
 
+// Structure for a single order item
 typedef struct {
-    char itemID[10];           // itemID stored as string for consistency
+    char itemID[10];
     char itemName[50];
     float price;
     int quantity;
-    char category[30]; 
-    char itemCategory[30];     // Needed for delivery time estimation
+    char itemCategory[30];  // For time estimation
 } Order;
 
-// Function declarations
+// Order handling functions
 int addToOrder(Order orders[], int* orderCount, const char itemID[], int quantity, const MenuItem menu[], int menuSize);
-int getMenuItemDetails(const MenuItem menu[], int menuSize, const char itemID[], char* itemName, float* price);
 void applyDiscount(Order orders[], int orderCount, char discountChoice);
+void printOrderSummary(Order orders[], int orderCount);
+
+// Time estimation
 int calculateDeliveryTime(Order orders[], int orderCount);
 int calculatePickupTime(Order orders[], int orderCount);
-void printOrderSummary(Order orders[], int orderCount);
