+++ Project II/order_flow.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void startCustomerOrder(const char* serviceType);
